+++ src/sardana/pool/poolcontrollers/test/test_DummyTriggerGateController.py
@@ -8,19 +8,19 @@ from sardana.pool.test import FakePool, createPoolController, \
     createPoolTriggerGate, dummyPoolTGCtrlConf01, dummyTriggerGateConf01, \
     createControllerConfiguration
 
-synchronization1 = [{SynchParam.Delay: {SynchDomain.Time: 0},
-                     SynchParam.Active: {SynchDomain.Time: .03},
-                     SynchParam.Total: {SynchDomain.Time: .1},
-                     SynchParam.Repeats: 0}]
+synch_description1 = [{SynchParam.Delay: {SynchDomain.Time: 0},
+                       SynchParam.Active: {SynchDomain.Time: .03},
+                       SynchParam.Total: {SynchDomain.Time: .1},
+                       SynchParam.Repeats: 0}]
 
-synchronization2 = [{SynchParam.Delay: {SynchDomain.Time: 0},
-                     SynchParam.Active: {SynchDomain.Time: .01},
-                     SynchParam.Total: {SynchDomain.Time: .02},
-                     SynchParam.Repeats: 10}]
+synch_description2 = [{SynchParam.Delay: {SynchDomain.Time: 0},
+                       SynchParam.Active: {SynchDomain.Time: .01},
+                       SynchParam.Total: {SynchDomain.Time: .02},
+                       SynchParam.Repeats: 10}]
 
 
-@insertTest(helper_name='generation', synchronization=synchronization1)
-@insertTest(helper_name='generation', synchronization=synchronization2)
+@insertTest(helper_name='generation', synch_description=synch_description1)
+@insertTest(helper_name='generation', synch_description=synch_description2)
 class PoolDummyTriggerGateTestCase(unittest.TestCase):
     """Parameterizable integration test of the PoolSynchronization action and
     the DummTriggerGateController.
@@ -49,10 +49,10 @@ class PoolDummyTriggerGateTestCase(unittest.TestCase):
         self.tg_action = PoolSynchronization(self.dummy_tg)
         self.tg_action.add_element(self.dummy_tg)
 
-    def generation(self, synchronization):
+    def generation(self, synch_description):
         """Verify that the created PoolTGAction start_action starts correctly
         the involved controller."""
-        args = ([self.ctrl_conf], synchronization)
+        args = ([self.ctrl_conf], synch_description)
         self.tg_action.start_action(*args)
         self.tg_action.action_loop()
         # TODO: add asserts applicable to a dummy controller e.g. listen to